(PC-6436) refacto of SearchResults and add condition on filter button rendering
