+++ src/features/search/components/InfiniteHits.tsx
@@ -4,9 +4,11 @@ import { FlatList } from 'react-native'
 import styled from 'styled-components/native'
 
 import { TAB_BAR_COMP_HEIGHT } from 'features/navigation/TabBar/TabBarComponent'
+import { FadeScrollingView } from 'features/search/atoms'
 import { Hit, NumberOfResults } from 'features/search/atoms'
+import { Filter } from 'features/search/atoms/Buttons'
 import { AlgoliaHit } from 'libs/algolia'
-import { ColorsEnum, getSpacing } from 'ui/theme'
+import { ColorsEnum, getSpacing, Spacer } from 'ui/theme'
 
 import { NoSearchResult } from '../atoms/NoSearchResult'
 
@@ -15,6 +17,7 @@ interface Props {
   hasMore: boolean
   refineNext: () => void
   handleIsScrolling: (value: boolean) => () => void
+  isScrolling: boolean
 }
 
 export const InfiniteHitsComponent: React.FC<Props> = ({
@@ -22,21 +25,32 @@ export const InfiniteHitsComponent: React.FC<Props> = ({
   hasMore,
   refineNext,
   handleIsScrolling,
+  isScrolling,
 }) => (
-  <Container>
-    <FlatList
-      data={hits}
-      keyExtractor={(item) => item.objectID}
-      ListHeaderComponent={NumberOfResults}
-      ListFooterComponent={Footer}
-      ItemSeparatorComponent={Separator}
-      renderItem={({ item: hit }) => <Hit hit={hit} />}
-      onEndReached={() => hasMore && refineNext()}
-      onScrollEndDrag={handleIsScrolling(false)}
-      onScrollBeginDrag={handleIsScrolling(true)}
-      ListEmptyComponent={NoSearchResult}
-    />
-  </Container>
+  <React.Fragment>
+    <Container>
+      <FlatList
+        data={hits}
+        keyExtractor={(item) => item.objectID}
+        ListHeaderComponent={NumberOfResults}
+        ListFooterComponent={Footer}
+        ItemSeparatorComponent={Separator}
+        renderItem={({ item: hit }) => <Hit hit={hit} />}
+        onEndReached={() => hasMore && refineNext()}
+        onScrollEndDrag={handleIsScrolling(false)}
+        onScrollBeginDrag={handleIsScrolling(true)}
+        ListEmptyComponent={NoSearchResult}
+      />
+    </Container>
+    {hits.length > 0 && (
+      <FilterContainer>
+        <FadeScrollingView isScrolling={isScrolling}>
+          <Filter />
+        </FadeScrollingView>
+        <Spacer.BottomScreen />
+      </FilterContainer>
+    )}
+  </React.Fragment>
 )
 
 const Container = styled.View({ height: '100%' })
@@ -46,6 +60,12 @@ const Separator = styled.View({
   backgroundColor: ColorsEnum.GREY_LIGHT,
   marginHorizontal: getSpacing(6),
   marginVertical: getSpacing(4),
+})
+
+const FilterContainer = styled.View({
+  alignSelf: 'center',
+  position: 'absolute',
+  bottom: TAB_BAR_COMP_HEIGHT + getSpacing(6),
 })
 
 export const InfiniteHits = connectInfiniteHits(InfiniteHitsComponent)
